requirements.txt
# PyTorch and related packages
torch>=2.0.0
torchvision>=0.15.0

# ONNX for model export and inference
onnx>=1.14.0
onnxruntime>=1.15.0

# Image processing and utilities
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Configuration and data handling
PyYAML>=6.0

# Progress bar
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# ROS2 Python packages (install via rosdep in ROS2 environment)
# rclpy
# sensor_msgs
# std_msgs

# Optional: for development
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
